+++ package/sdk/src/lib/security.ts
@@ -68,13 +68,20 @@ type SecurityInputOAuth2ClientCredentials = {
     value: { clientID?: string | undefined; clientSecret?: string | undefined } | null | undefined;
 };
 
+type SecurityInputCustom = {
+    type: "http:custom";
+    value: any | null | undefined;
+    fieldName: string;
+};
+
 export type SecurityInput =
     | SecurityInputBasic
     | SecurityInputBearer
     | SecurityInputAPIKey
     | SecurityInputOAuth2
     | SecurityInputOAuth2ClientCredentials
-    | SecurityInputOIDC;
+    | SecurityInputOIDC
+    | SecurityInputCustom;
 
 export function resolveSecurity(...options: SecurityInput[][]): SecurityState | null {
     const state: SecurityState = {
@@ -90,6 +97,8 @@ export function resolveSecurity(...options: SecurityInput[][]): SecurityState | 
                 return false;
             } else if (o.type === "http:basic") {
                 return o.value.username != null || o.value.password != null;
+            } else if (o.type === "http:custom") {
+                return null;
             } else if (o.type === "oauth2:client_credentials") {
                 return o.value.clientID != null || o.value.clientSecret != null;
             } else if (typeof o.value === "string") {
@@ -124,6 +133,8 @@ export function resolveSecurity(...options: SecurityInput[][]): SecurityState | 
                 break;
             case "http:basic":
                 applyBasic(state, spec);
+                break;
+            case "http:custom":
                 break;
             case "http:bearer":
                 applyBearer(state, spec);